Added some text to the processing system overview
Summary:
Added text to tprocessing system overview (Chapter 2).
The document compiles, but with problems:
    - The amsmath package causes a conflict, probably due to some incompatible
      package
    - NatBib complains about undefined references for a reason I cannot at the
      moment fathom
Both these issues need to be resolved, but are beyond the scope (at the grace
of the right-honourable reviewrs of course!) of the current changeset.

Test Plan:
Text is sensible. PDF generated. Though in the opposite order of
priority.

Reviewers: bengt.rydberg, jimmy

Reviewed By: jimmy

Subscribers: skymandr

Differential Revision: http://phabricator.molflow.com/D64